pyproject.toml
[project]
name = "uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "bcrypt>=4.2.0",
    "dotenv>=0.9.9",
    "email-validator>=2.1.0",
    "fastapi>=0.128.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "PyJWT>=2.9.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
